configure: automatically parse command line for meson -D options
Right now meson_options.txt lists about 50 options.  Each option
needs code in configure to parse it and pass the option down to Meson as
a -D command-line argument; in addition the default must be duplicated
between configure and meson_options.txt.

This series tries to remove the code duplication by passing unknown
--enable and --disable options to a Perl program, and using Meson's
introspection support to match those to meson_options.txt.
About 75% of the options can be handled completely by the new
mechanism; nine need to be parsed in configure but still have
their help automatically generated.

The main advantage is less code in configure and a more consistent
parsing and help.  For example --enable-slirp and --enable-blobs
were not supported.

The disadvantages are:

- a few options change name: --enable-tcmalloc and --enable-jemalloc
become --enable-malloc={tcmalloc,jemalloc}; --enable-blobs becomes
--enable-install-blobs.

- because we parse command-line options before meson is available,
the introspection output is stored in the source tree.  The output
is slightly modified using the "jq" tool in order to ensure that it's
stable and that modifications to meson_buildoptions.txt do not cause
horrible conflicts.  This is the main reason for the unattractive
diffstat (the number of JSON lines added is higher than the number
of configure lines removed, though of course the latter are code
that must be maintained manually and the former is not).

- because we need to run the script to generate the full help, we
cannot rely on the user supplying the path to a Python interpreter
with --python.  For this reason, the script is written in Perl.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>