meson: remove config_all
Meson sourcesets have all_sources() and all_dependencies() method that
automatically gather both the if_true and the if_false side of the
rules.  This removes the need for the ugly and error-prone CONFIG_ALL
symbol.

The downside of using all_sources() is that build.ninja and
compile_commands.json will also include rules to build
target-dependent files that have been compiled out for this
build.  However these rules are not used and the extra
files will not be built by a normal "make".

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>