curl: convert to meson, removing curl-config and requiring at least 7.29.0
pkg-config is enough for QEMU, since it does not need extra information
such as the path for certicate authorities.  All supported platforms
today will all have pkg-config for curl, so we can drop curl-config.

QEMU has compatibility code for curl versions older than 7.16.0, but it
is in all likelihood untested (7.16.0 was released in 2006).  Require a
minimum version in preparation for removing that code; as suggested by
Daniel Berrangé, I am using the oldest version found in currently
supported distros, which is 7.29.0 from CentOS 7.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>